+++ packages/components/modules/__shared__/common/types.ts
@@ -2,6 +2,7 @@ import { ReactNode } from 'react'
 
 export type SocialUpsertForm = {
   body: string
+  mentionedProfileIds: string[]
 }
 
 export interface SearchNotFoundStateProps {