+++ packages/components/modules/comments/web/CommentCreate/types.ts
@@ -1,5 +1,7 @@
 import { FC } from 'react'
 
+import type { MentionsSearchController } from '@baseapp-frontend/design-system/components/web/inputs'
+
 import type { SocialInputProps } from '../../../__shared__/web'
 
 export interface CommentCreateProps {
@@ -7,4 +9,6 @@ export interface CommentCreateProps {
   autoFocusInput?: boolean
   SocialInput?: FC<SocialInputProps>
   SocialInputProps?: Partial<SocialInputProps>
+  mentionsController?: MentionsSearchController
+  disableMentions?: boolean
 }