+++ packages/design-system/components/web/inputs/MarkdownEditorField/InitializedMDXEditor/types.ts
@@ -3,7 +3,9 @@ import { ForwardedRef } from 'react'
 import type { MDXEditorMethods } from '@mdxeditor/editor'
 
 import { MarkdownEditorProps } from '../types'
+import type { MentionsActiveConfig } from './plugins/mentions/types'
 
 export type InitializedMDXEditorProps = {
   editorRef: ForwardedRef<MDXEditorMethods> | null
-} & MarkdownEditorProps
+  mentions?: MentionsActiveConfig
+} & Omit<MarkdownEditorProps, 'mentions'>